+++ src/app/pages/Home.tsx
@@ -1,4 +1,4 @@
-import Header from '../components/Header'
+import Header from '../../components/Header'
 
 const Home = () => {
     const user = false
