+++ src/app/pages/Search.tsx
@@ -1,4 +1,4 @@
-import Header from '../components/Header'
+import Header from '../../components/Header'
 
 const Search = () => {
     return (
