+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 
 import './main.css'
-import App from './App'
+import App from './app/App'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
